+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Sheet coverage added to give an approximate of how much of the sheet has been filled out
 - Documents and Links can now have their categories renamed *en masse*
 - Project Editors, Viewers, and Site Members can now comment on sheets
+- Project news posts are now emailed to project editors, viewers, and site members when they are created
+  - Archived news posts are not sent out via email
 
 ### Bug Fix
 - `Include missing` is now always displayed on the design report